pyproject.toml
[project]
name = "agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "python-dotenv>=1.0.1",
    "boto3>=1.35.53",
    "duckdb>=1.2.2",
    "duckduckgo-search>=8.0.0",
    "fastapi>=0.115.12",
    "qdrant-client>=1.13.3",
    "spacy>=3.8.5",
    "pydantic[email]>=2.10.6",
    "pydantic-ai>=0.0.40",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
    "matplotlib>=3.10.1",
    "fastapi-mcp>=0.2.0",
    "requests>=2.32.3",
    "tika>=3.1.0",
]
